requirements.txt
click==7.1.2
environs==9.3.1
joblib==1.0.1
marshmallow==3.10.0
mysql-connector-python==8.0.23
nltk==3.5
numpy==1.20.1
pkg-resources==0.0.0
protobuf==3.14.0
python-dotenv==0.15.0
regex==2020.11.13
six==1.15.0
sner==0.2.12
tqdm==4.56.2
